Reduce remaining complexity findings on alerts overlay code
Splits alertsAdminMutations.ts into three files by concern —
alertsShared.ts (parseEventType/NOT_A_STREAMER_REDIRECT, used by both),
alertsAssetMutations.ts (image/sound upload+delete routes), and a slimmed
alertsAdminMutations.ts (config-save + test-alert routes only) — since
the "high total complexity" finding is a per-file total and the file had
grown to cover two largely independent concerns. Also breaks
createSseEventsHandler's keepalive/cleanup logic out into small
removeClient/sendSseHandshake/startKeepalive helpers to bring its own
function-level complexity down.
