+++ src/web/routes/alertsAdmin.ts
@@ -6,7 +6,8 @@ import { getSessionUser } from '../session';
 import { getStreamerByDiscordId, getAlertConfigsForStreamer, ALERT_EVENT_TYPES } from '../../db';
 import { PUBLIC_URL } from '../../shared/config';
 import { filterQueryParam, renderError, renderView } from './shared';
-import { router as mutationsRouter, MAX_IMAGE_MB, MAX_SOUND_MB } from './alertsAdminMutations';
+import { router as mutationsRouter } from './alertsAdminMutations';
+import { router as assetMutationsRouter, MAX_IMAGE_MB, MAX_SOUND_MB } from './alertsAssetMutations';
 
 const log = createLogger('AlertsAdmin');
 const router = Router();
@@ -52,5 +53,6 @@ router.get('/settings', requireAuth, csrfProtection, async (req, res) => {
 });
 
 router.use(mutationsRouter);
+router.use(assetMutationsRouter);
 
 export default router;
