+++ src/web/routes/sseChannel.ts
@@ -1,5 +1,7 @@
 import type { Request, Response, NextFunction } from 'express';
 
+const KEEPALIVE_INTERVAL_MS = 25_000;
+
 /** Options for {@link createSseEventsHandler}. */
 export interface SseEventsHandlerOptions {
   /** In-memory map of active SSE connections keyed by lowercased channel login. */
@@ -12,11 +14,47 @@ export interface SseEventsHandlerOptions {
   maxPerChannel: number;
 }
 
+/** Removes `res` from the channel's client Set, deleting the map entry once it's empty. */
+function removeClient(connections: Map<string, Set<Response>>, key: string, res: Response): void {
+  const clients = connections.get(key);
+  if (!clients) return;
+  clients.delete(res);
+  if (clients.size === 0) connections.delete(key);
+}
+
+/** Writes the SSE handshake headers and the initial `: connected` comment. */
+function sendSseHandshake(res: Response): void {
+  res.setHeader('Content-Type', 'text/event-stream');
+  res.setHeader('Cache-Control', 'no-cache');
+  res.setHeader('Connection', 'keep-alive');
+  res.setHeader('X-Accel-Buffering', 'no'); // disable Nginx buffering if behind proxy
+  res.flushHeaders();
+  res.write(': connected\n\n');
+}
+
+/**
+ * Starts the periodic keepalive ping for one connection. If a ping write fails (e.g. the client
+ * disconnected without the request's 'close' event firing first), evicts the client and clears
+ * the interval so nothing is left running for a dead connection.
+ * @returns The interval handle, so the caller can clear it on a normal 'close' event too.
+ */
+function startKeepalive(connections: Map<string, Set<Response>>, key: string, res: Response): NodeJS.Timeout {
+  const keepalive = setInterval(() => {
+    try {
+      res.write(': ping\n\n');
+    } catch {
+      clearInterval(keepalive);
+      removeClient(connections, key, res);
+    }
+  }, KEEPALIVE_INTERVAL_MS);
+  return keepalive;
+}
+
 /**
  * Builds a `/:login/events`-style SSE route handler, shared by the reward-video overlay and the
  * alerts overlay (each keeps its own `connections` map and push function, since those differ in
  * payload shape — this only factors out the identical connection lifecycle: login validation,
- * per-channel connection-limit enforcement, SSE handshake headers, the 25s keepalive ping, and
+ * per-channel connection-limit enforcement, SSE handshake headers, the keepalive ping, and
  * cleanup on a failed write or client disconnect).
  * @param options - See {@link SseEventsHandlerOptions}.
  * @returns An Express route handler: on a valid, non-reserved login, upgrades the response to
@@ -42,33 +80,12 @@ export function createSseEventsHandler(
       return;
     }
 
-    res.setHeader('Content-Type', 'text/event-stream');
-    res.setHeader('Cache-Control', 'no-cache');
-    res.setHeader('Connection', 'keep-alive');
-    res.setHeader('X-Accel-Buffering', 'no'); // disable Nginx buffering if behind proxy
-    res.flushHeaders();
-
-    res.write(': connected\n\n');
-
-    const keepalive = setInterval(() => {
-      try {
-        res.write(': ping\n\n');
-      } catch {
-        clearInterval(keepalive);
-        const clients = connections.get(key);
-        if (clients) {
-          clients.delete(res);
-          if (clients.size === 0) connections.delete(key);
-        }
-      }
-    }, 25_000);
+    sendSseHandshake(res);
+    const keepalive = startKeepalive(connections, key, res);
 
     req.on('close', () => {
       clearInterval(keepalive);
-      const clients = connections.get(key);
-      if (!clients) return;
-      clients.delete(res);
-      if (clients.size === 0) connections.delete(key);
+      removeClient(connections, key, res);
     });
   };
 }
